refactor: Cards.tsx 코드 리팩토링
